fix load_stack when dimension metadata is incorrectly stored
I have some data (nx=512,ny=256) where the normal dimension metadata is incorrectly listed as ny=512, but readlif correctly finds the image size. This is strange and probably breaks other things, but at least this avoids an error in load_stack reshaping and it is anyway more performant than before.